+++ packages/amplify_auth_cognito/android/src/main/kotlin/com/amazonaws/amplify/amplify_auth_cognito/types/FlutterSignInRequest.kt
@@ -15,15 +15,26 @@
 
 package com.amazonaws.amplify.amplify_auth_cognito.types
 
+import androidx.annotation.NonNull
 import com.amazonaws.amplify.amplify_core.exception.ExceptionMessages
 import com.amazonaws.amplify.amplify_core.exception.InvalidRequestException
 import com.amplifyframework.auth.AuthUser
+import com.amplifyframework.auth.cognito.options.AWSCognitoAuthSignInOptions
 import com.amplifyframework.core.Amplify
 
 data class FlutterSignInRequest(val map: HashMap<String, *>) {
   val username: String = map["username"] as String;
   val password: String = map["password"] as String;
-  val options: HashMap<String, *>? = map["options"] as HashMap<String, *>?;
+  val options: AWSCognitoAuthSignInOptions = formatOptions(map["options"] as HashMap<String, *>?)
+
+  private fun formatOptions(rawOptions: HashMap<String, *>?): AWSCognitoAuthSignInOptions {
+    var options =  AWSCognitoAuthSignInOptions.builder();
+
+    if(rawOptions?.get("clientMetadata") != null)
+      options.metadata(rawOptions["clientMetadata"] as HashMap<String, String>);
+
+    return options.build();
+  }
 
   companion object {
     private const val validationErrorMessage: String = "SignIn Request malformed."